Refactored to use a base class that is inheritted byt the pages
