+++ angular/src/app/helpers/dhis-link-constants.ts
@@ -4,8 +4,8 @@ export const REFRESH_TOKEN = 'REFRESH_TOKEN';
 export const FORM_DATA = 'FORM_DATA';
 export const CURRENT_ROUTE = 'CURRENT_ROUTE';
 export const CURRENT_USER = 'CURRENT_USER';
-//export const BASE_URL = 'https://ehealth.ub.bw/api/ddpcontroller/';
+export const BASE_URL = 'https://ehealth.ub.bw/api/ddpcontroller/';
 
-export const BASE_URL = 'http://localhost:8080/ddpcontroller/';
+//export const BASE_URL = 'http://localhost:8080/ddpcontroller/';
 
 export class DhisLinkConstants {}
