+++ angular/src/app/view/batch/batch.component.html
@@ -1,0 +1,1 @@
+
